Implement score based reward system
Generally it's still the same proportional system, but uses exponential share weighting.

The main idea is a distribution of rewards according to proportion of miner's score in the total score. Score is calculated using this simple formula:

S = share_diff * exp(cumulative_diff / network_diff)

Where cumulative_diff is a sum of difficulty for all shares of the current round.